+++ test/unit/mysql_perf/code_coverage.sh
@@ -11,11 +11,14 @@ echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mysql_perf test/unit/mysql_perf/create_data_config.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/create_header.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/data_out.py
+coverage run -a --source=mysql_perf test/unit/mysql_perf/file_out.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/help_message.py
+coverage run -a --source=mysql_perf test/unit/mysql_perf/mail_out.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/main.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/mysql_stat.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/mysql_stat_run.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/run_program.py
+coverage run -a --source=mysql_perf test/unit/mysql_perf/std_out.py
 
 echo ""
 echo "Producing code coverage report"